新增個區塊component與基本樣式
