+++ assets/css/variables.css
@@ -58,6 +58,7 @@
   --main-bg: var(--color-primary-light);
   --filter-bg: var(--color-white);
   --filter-item-bg: var(--color-primary-main);
+  --block-title-bg: var(--color-primary-main);
   --footer-bg: var(--color-primary-main);
   --copyright-bg: var(--color-black);
   --border-color: var(--color-gray-200);
@@ -73,6 +74,7 @@
   --main-bg: var(--color-gray-700);
   --filter-bg: var(--color-primary-main);
   --filter-item-bg: var(--color-white);
+  --block-title-bg: var(--color-black);
   --footer-bg: var(--color-black);
   --copyright-bg: var(--color-gray-700);
   --border-color: var(--color-gray-700);
