--- components/Card/Button.vue
@@ -1,7 +1,0 @@
-<template>
-  <button>click</button>
-</template>
-
-<script setup></script>
-
-<style scoped></style>
