--- components/Card/index.vue
@@ -1,22 +1,0 @@
-<template>
-  <div class="card">
-    <span>Component card</span><br />
-    <Button />
-  </div>
-</template>
-
-<script setup>
-import Button from "@/components/Card/Button.vue";
-</script>
-
-<style scoped>
-.card {
-  width: 100%;
-  height: 240px;
-  background: rgba(211, 211, 211, 0.4);
-  color: var(--color-text);
-  border-radius: 0.5rem;
-  text-align: center;
-  padding-top: 5rem;
-}
-</style>
