+++ components/Policy/index.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="policy">
+    <Title />
+    <button>click</button>
+  </div>
+</template>
+
+<script setup>
+import Title from "@/components/Policy/Title.vue";
+</script>
+
+<style scoped>
+.policy {
+  width: 100%;
+  /* height: auto; */
+  height: 200px;
+}
+</style>
