+++ components/Sponsor/index.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="sponsor">
+    <Title />
+    <button>click</button>
+  </div>
+</template>
+
+<script setup>
+import Title from "@/components/Sponsor/Title.vue";
+</script>
+
+<style scoped>
+.sponsor {
+  width: 100%;
+  /* height: auto; */
+  height: 200px;
+}
+</style>
