+++ pages/index.vue
@@ -1,12 +1,12 @@
 <template>
   <div class="page">
     <Loading />
-    <Card />
     <Introduce />
-    <IntroduceBox />
-    <h2>喵喵的最新軌跡</h2>
+    <!-- <IntroduceBox /> -->
+    <Policy />
     <Timeline />
-    <ContactForm />
+    <Sponsor />
+    <Contact />
   </div>
 </template>
