--- pages/timelineTest.vue
@@ -1,6 +1,0 @@
-<template>
-  <div>
-    <Timeline></Timeline>
-  </div>
-</template>
-
